Extra files pulled, deleted/reorganized them
