+++ client/src/views/CreateEvent.tsx
@@ -1,5 +1,4 @@
 import React from 'react'
-
 import axios from 'axios'
 import { useNavigate } from 'react-router-dom'
 import EventForm from '../components/forms/EventForm'
