--- client/src/views/EventDisplay.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import Display from '../components/Display'
-import HeaderBar from '../components/HeaderBar'
-
-const EventDisplay = () => {
-    return (<>
-    <HeaderBar />
-    <Display />
-    </>)
-}
-
-export default EventDisplay
