--- client/src/views/EventPage 2.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import HeaderBar from '../components/HeaderBar'
-import EventDisplay from './EventDisplay'
-
-const EventPage = () => {
-    return (<>
-        <HeaderBar title='{event.name}' btnTitle='Dashboard' btnRoute='dashboard'/>
-        <EventDisplay />
-    </>)
-}
-
-export default EventPage
